requirements.txt

qiskit>=2.1.2
qiskit-aer>=0.17.1
mqt.ddsim>=2.0.0
stim>=1.15.0
numpy
mqt-core>=3.2.1
